Create Naughty or Nice
The const month in the for (const month in data) loop is a variable that I created to iterate over the keys of the data object, which in this case are the names of the months. In each iteration of the loop, month will hold the name of the current month being processed.

The for...in loop is a way to iterate over the properties of an object in JavaScript. In this context, month is a temporary variable that takes on the value of each key in the data object as the loop iterates. It's similar to how you might use a variable in a traditional for loop to hold the value of the current index.